Add option to specify the data folder
Specifying the data folder is optional. No functional difference is introduced when using the absolute paths as in the task JSON files. However, this adds the option to specify a relative path in the JSON files, and specify the base path in the command. This is useful when working on various systems with different file structure.